requirements.txt
Flask>=2.3
Pillow>=10.3
numpy>=1.26
gunicorn>=21.2
stegoveritas>=1.10
stegcracker>=2.0.9
volatility3>=2.5.2
pytest>=7.4
stegpy>=0.0.1
stego-lsb>=1.4
stegano>=0.11
jpeglib>=1.0
soundfile>=0.12
scipy>=1.11
